LayoutTests/fast/profiler/anonymous-function-called-from-different-contexts.html

<html>
<head>
<script src="resources/profiler-test-JS-resources.js"></script>
<script>
if (window.testRunner)
    testRunner.dumpAsText();
if (window.internals)
    internals.setLegacyJavaScriptProfilingEnabled(true);

function startTest()
{
    console.profile("Same anonymous function called from different contexts");

    anonymousFunction();
    eval("anonymousFunction()");

    endTest();
}
</script>
</head>

<body onload="startTest()">
This page's JavaScript calls an anonymous function from different contexts.
<br>
<br>
To run this test manually, load it in the browser then load the WebInspector and look at
the profile.  Beneath onload and startTest it should show three children, an
anonymous function, script context, and the endTest call.
<div id="output"></div>
</body>
</html>
