LayoutTests/fast/profiler/anonymous-function-called-from-different-contexts-expected.txt

This page's JavaScript calls an anonymous function from different contexts. 

To run this test manually, load it in the browser then load the WebInspector and look at the profile. Beneath onload and startTest it should show three children, an anonymous function, script context, and the endTest call.

Profile title: Same anonymous function called from different contexts
Thread_1 (no file) (line 0)
   startTest anonymous-function-called-from-different-contexts.html (line 11)
      anonymousFunction profiler-test-JS-resources.js (line 29)
         insertNewText profiler-test-JS-resources.js (line 17)
            createElement (no file) (line 0)
            createTextNode (no file) (line 0)
            appendChild (no file) (line 0)
            getElementById (no file) (line 0)
      eval (no file) (line 0)
         (program) (no file) (line 1)
            anonymousFunction profiler-test-JS-resources.js (line 29)
               insertNewText profiler-test-JS-resources.js (line 17)
                  createElement (no file) (line 0)
                  createTextNode (no file) (line 0)
                  appendChild (no file) (line 0)
                  getElementById (no file) (line 0)
      endTest profiler-test-JS-resources.js (line 1)
